tried to work on packagejson and testing setup and more json object
