--- jsonObject/json.css
@@ -1,20 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Open+Sans, Lato);
-
-body {
-    font-family: 'Open Sans', sans-serif;
-    font-size: 36px;
-    color: #000;
-    background-image: url("snow.jpg");
-    no-repeat center center fixed;
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-    position: cover;
-    padding-top: 30px;
-    padding-left: 250px;
-    -webkit-transition: all 12s ease-in-out;
-    -moz-transition: all 12s ease-in-out;
-    -o-transition: all 12s ease-in-out;
-    transition: all 5s ease-in-out;
-}
