--- jsonObject/json.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<title>JSON object</title>
-	<link rel="stylesheet" type="text/css" href="json.css">
-</head>
-
-<body>
-	<script src="json.js" type="text/javascript"></script>
-</body>
-
-</html>
