--- scrab-bag/scrab.js
@@ -1,4 +1,0 @@
-'use strict'
-const tiles = require('./tiles')
-
-let pickedTiles =
