+++ src/json.js
@@ -1,0 +1,22 @@
+var jsonStrings = {
+    "name": "William Shakespeare",
+    "dead": true,
+    "works": [{
+            "name": "Romeo and Juliet",
+            "published": 1591,
+            "isAwesome": true
+        },
+        {
+            "name": "Richard III",
+            "published": 1592,
+            "isAwesome": false
+        }
+    ],
+    "favoriteSites": [
+        "tumblr",
+        "4chan"
+    ]
+};
+
+var myJSON = JSON.stringify(jsonStrings);
+alert(myJSON);
