+++ src/scrab.js
@@ -1,0 +1,6 @@
+
+const scrab = string => {
+
+  
+
+export default scrab
