+++ test/scrab.js
@@ -1,0 +1,15 @@
+const chai = require('chai')
+const expect = chai.expect
+import scrab from '../src/scrab.js'
+
+describe('scrab', () => {
+  it('', () => {
+    expect()
+  })
+  it('', () => {
+    expect()
+  })
+  it('', () => {
+    expect()
+  })
+})
